requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# ML/AI
torch>=2.2.0
transformers>=4.35.2
sentence-transformers>=2.2.2
numpy>=1.24.3
faiss-cpu>=1.7.4

# Vector DB alternatives
chromadb==0.4.18

# Utilities
python-dotenv==1.0.0
requests==2.31.0
